Adhere to the proper collector interface
